Change from dynamic allocation to static in graphx.c
The compiler can't detect how much memory dynamic allocation would be needed
on the microcontroller. This means that the real memory usage is not
displayed after compiling and I have caused a usage more than 100%.

Because the size of the display buffer is always the same (during
runtime), a static memory allocation is used instead (the buffer is
represented as a global static variable).